removed modules/common and lint+formatted code
