+++ src/components/planner/Tiles/SemesterCourseItem.tsx
@@ -2,7 +2,7 @@ import { UniqueIdentifier, useDraggable } from '@dnd-kit/core';
 import CloseIcon from '@mui/icons-material/Close';
 import { ComponentPropsWithoutRef, FC, forwardRef } from 'react';
 
-import { getFirstNewSemester,isEarlierSemester } from '@/utils/plannerUtils';
+import { getFirstNewSemester, isEarlierSemester } from '@/utils/plannerUtils';
 
 import { DragDataFromSemesterTile, DraggableCourse, Semester } from '../types';
