+++ src/components/planner/Tiles/SemesterTile.tsx
@@ -1,8 +1,7 @@
 import { UniqueIdentifier, useDroppable } from '@dnd-kit/core';
-import { SemesterCode } from '@prisma/client';
 import { FC, forwardRef } from 'react';
 
-import { getFirstNewSemester,isEarlierSemester } from '@/utils/plannerUtils';
+import { getFirstNewSemester, isEarlierSemester } from '@/utils/plannerUtils';
 
 import {
   DragDataToSemesterTile,
@@ -19,17 +18,6 @@ export interface SemesterTileProps {
   isValid: boolean;
   isDisabled: boolean;
   onRemoveCourse: (semester: Semester, course: DraggableCourse) => void;
-}
-
-export function displaySemesterCode(semesterCode: SemesterCode): string {
-  let semesterName;
-  if (semesterCode.semester === 'f') {
-    semesterName = 'Fall';
-  } else if (semesterCode.semester === 's') semesterName = 'Spring';
-  else {
-    semesterName = 'Summer';
-  }
-  return `${semesterName} ${semesterCode.year}`;
 }
 
 function getTitleText({ isValid, isDisabled }: { isValid: boolean; isDisabled: boolean }) {
