+++ src/pages/app/plans/[planId].tsx
@@ -1,13 +1,16 @@
 import { SemesterCode } from '@prisma/client';
+import { PDFDownloadLink } from '@react-pdf/renderer';
 import { createProxySSGHelpers } from '@trpc/react-query/ssg';
 import { ObjectID } from 'bson';
 import { useRouter } from 'next/router';
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next/types';
 import { unstable_getServerSession } from 'next-auth';
+import { useSession } from 'next-auth/react';
 import { useState } from 'react';
 import React from 'react';
 import superjson from 'superjson';
 
+import DegreePlanPDF from '@/components/planner/DegreePlanPDF/DegreePlanPDF';
 import Planner from '@/components/planner/Planner';
 import { DraggableCourse, Semester, ToastMessage } from '@/components/planner/types';
 import BackArrowIcon from '@/icons/BackArrowIcon';
@@ -18,9 +21,6 @@ import { appRouter } from '@/server/trpc/router/_app';
 import { trpc } from '@/utils/trpc';
 import { useTaskQueue } from '@/utils/useTaskQueue';
 import { createNewYear } from '@/utils/utilFunctions';
-import { PDFDownloadLink } from '@react-pdf/renderer';
-import DegreePlanPDF from '@/components/planner/DegreePlanPDF/DegreePlanPDF';
-import { useSession } from 'next-auth/react';
 
 /**
  * A page that displays the details of a specific student academic plan.
